requirements.txt
networkx
matplotlib.pyplot
numpy 
scipy